move test into dedicated test file
